Permanently deleting a subscriber seems wrong
I've rewritten the delete section slightly as it was a bit confusing.

But mainly the first code example didn't exist in Mailcoach, as far as I can see - so I removed it.

The very last line isn't very clear either. Seems to be that the code example that's left just gets the Subscriber model and calls delete, which is an Eloquent method. The only thing I can think this refers to is that by deleting a Subscriber, you will also be deleting all their stats (due tot he cascades defined in the migration file).

I'm not sure what to update that last line to - so have left as is...