Provide clarity on how attributes are set upon creation
These docs are not clear that attributes are set based on the config values at the time of EmailList creation. This caused me a lot of trouble, so I'm hoping this docs change provides clarity for others.

To me, it would be more intuitive for the app to use the config values on the EmailList and leave the attributes set to `null`. I would only set this value if the user chose to override the default. By setting these values on the creation of every EmailList, that means if have a large number of lists and want to change my mailer for all of them, I have to change the attribute on each list instead of just changing my config value.